Replaced front-end part
