+++ html_output/public/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="initial-scale=0.1">
+    <meta name="author" content="Khalid El Haji">
+    <title>Geit | Git educational insights tool</title>
+    <script src="https://kit.fontawesome.com/995ad4888d.js"></script>
+  </head>
+  <body  class="bg-light">
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root"></div>
+  </body>
+</html>
