--- webpage/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "git-edu-insights-webpage-output",
-  "version": "1.0.0",
-  "description": "The HTML front for the insights generated by the tool.",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "watch": "./node_modules/.bin/webpack -d --watch",
-    "minimize": "./node_modules/.bin/webpack --mode production --optimize-minimize",
-    "run": "./node_modules/.bin/webpack -d"
-  },
-  "author": "Khalid El Haji",
-  "license": "MIT",
-  "dependencies": {
-    "moment": "^2.27.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "recharts": "^1.8.5"
-  },
-  "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.5",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "html-webpack-inline-source-plugin": "0.0.10",
-    "html-webpack-plugin": "^3.2.0",
-    "terser-webpack-plugin": "^2.3.8",
-    "webpack": "^4.44.1",
-    "webpack-bundle-analyzer": "^3.8.0",
-    "webpack-cli": "^3.3.12"
-  }
-}
